removed dependency setup from bg jobs and move neccesary code and similarly for app_config.json and updated imports
