--- byoeb-v1/byoeb/byoeb/background_jobs/dependency_setup.py
@@ -1,177 +1,0 @@
-from byoeb.factory import (
-    ChannelClientFactory,
-    MongoDBFactory
-)
-import byoeb.background_jobs.config as env_config
-from byoeb.background_jobs.config import app_config
-from byoeb.services.databases.mongo_db import UserMongoDBService, MessageMongoDBService
-
-
-
-SINGLETON = "singleton"
-
-# channel
-channel_client_factory = ChannelClientFactory(config=app_config)
-
-# mongo db
-mongo_db_factory = MongoDBFactory(
-    config=app_config,
-    scope=SINGLETON
-)
-
-from byoeb.services.leaderboard import LeaderboardService
-from typing import Optional
-
-# Initialize MongoDB services
-_user_db_service = UserMongoDBService(
-    config=app_config,
-    mongo_db_factory=mongo_db_factory
-)
-_message_db_service = MessageMongoDBService(
-    config=app_config,
-    mongo_db_factory=mongo_db_factory,
-    user_db_service=_user_db_service  # Pass user_db_service for leaderboard functionality
-)
-
-# Service instances for leaderboard (lazy initialization)
-_leaderboard_service: Optional[LeaderboardService] = None
-
-async def get_user_service() -> UserMongoDBService:
-    """Get user MongoDB service instance."""
-    return _user_db_service
-
-async def get_message_service() -> MessageMongoDBService:
-    """Get message MongoDB service instance."""
-    return _message_db_service
-
-async def get_leaderboard_service() -> LeaderboardService:
-    """Get or create leaderboard service instance."""
-    global _leaderboard_service
-    if _leaderboard_service is None:
-        user_service = await get_user_service()
-        message_service = await get_message_service()
-        _leaderboard_service = LeaderboardService(user_service, message_service)
-    return _leaderboard_service
-
-# Keep backward compatibility - expose these as module-level variables
-user_db_service = _user_db_service
-message_db_service = _message_db_service
-
-# Text translator
-from byoeb_integrations.translators.text.azure.async_azure_text_translator import AsyncAzureTextTranslator
-from azure.identity import get_bearer_token_provider, AzureCliCredential
-
-token_provider = get_bearer_token_provider(
-    AzureCliCredential(), app_config["app"]["azure_cognitive_endpoint"]
-)
-# TODO: factory implementation
-text_translator = AsyncAzureTextTranslator(
-    credential=AzureCliCredential(),
-    region=app_config["translators"]["text"]["azure_cognitive"]["region"],
-    resource_id=app_config["translators"]["text"]["azure_cognitive"]["resource_id"],
-)
-
-# Speech translator
-# TODO: factory implementation
-from byoeb_integrations.translators.speech.azure.async_azure_speech_translator import AsyncAzureSpeechTranslator
-voice_dict = {
-    "male": {
-        "en-IN": "en-IN-PrabhatNeural",
-        "hi-IN": "hi-IN-MadhurNeural",
-        "mr-IN": "mr-IN-ManoharNeural"
-    },
-    "female": {
-        "en-IN": "en-IN-NeerjaNeural",
-        "hi-IN": "hi-IN-SwaraNeural",
-        "mr-IN": "mr-IN-AarohiNeural"
-    }
-}
-
-speech_translator = AsyncAzureSpeechTranslator(
-    token_provider=token_provider,
-    region=app_config["translators"]["speech"]["azure_cognitive"]["region"],
-    resource_id=app_config["translators"]["speech"]["azure_cognitive"]["resource_id"],
-)
-speech_translator.change_voice_dict(voice_dict)
-
-from byoeb_integrations.translators.speech.azure.async_azure_openai_whisper import AsyncAzureOpenAIWhisper
-speech_translator_whisper = AsyncAzureOpenAIWhisper(
-    token_provider=token_provider,
-    model=app_config["translators"]["speech"]["azure_oai"]["model"],
-    azure_endpoint=app_config["translators"]["speech"]["azure_oai"]["endpoint"],
-    api_version=app_config["translators"]["speech"]["azure_oai"]["api_version"]
-)
-
-# llm
-# from byoeb_integrations.llms.llama_index.llama_index_azure_openai import AsyncLLamaIndexAzureOpenAILLM
-# llm_client = AsyncLLamaIndexAzureOpenAILLM(
-#     model=app_config["llms"]["azure"]["model"],
-#     deployment_name=app_config["llms"]["azure"]["deployment_name"],
-#     azure_endpoint=app_config["llms"]["azure"]["endpoint"],
-#     token_provider=token_provider,
-#     api_version=app_config["llms"]["azure"]["api_version"]
-# )
-from byoeb_integrations.llms.llama_index.llama_index_openai import AsyncLLamaIndexOpenAILLM
-llm_client = AsyncLLamaIndexOpenAILLM(
-    model=app_config["llms"]["openai"]["model"],
-    api_key=env_config.env_openai_api_key,
-    api_version=app_config["llms"]["openai"]["api_version"],
-    organization=env_config.env_openai_org_id
-)
-
-from byoeb.application_logger.azure_app_insights import AzureAppInsightsLogger
-app_insights_logger = None
-if env_config.env_appinsights_connection_string:
-    print("✅ App Insights connection string set. Enabling Azure logging.")
-    app_insights_logger = AzureAppInsightsLogger(
-        logger_name=app_config["app_logger"]["azure"]["logger_name"],
-        connection_string=env_config.env_appinsights_connection_string,
-        instrumentations=["fastapi", "urllib3"]
-    )
-else:
-    print("⚠️ App Insights connection string not set. Skipping Azure logging.")
-
-# Scheduler configuration
-from apscheduler.schedulers.asyncio import AsyncIOScheduler
-from apscheduler.triggers.cron import CronTrigger
-from apscheduler.jobstores.mongodb import MongoDBJobStore
-from apscheduler.executors.asyncio import AsyncIOExecutor
-from apscheduler.events import EVENT_JOB_EXECUTED, EVENT_JOB_ERROR
-import pymongo
-from byoeb.chat_app.configuration.config import env_mongo_db_connection_string
-
-# MongoDB connection configuration for scheduler job store
-MONGODB_URL = env_mongo_db_connection_string
-MONGODB_DATABASE = app_config["databases"]["mongo_db"]["database_name"]
-MONGODB_COLLECTION = app_config["databases"]["mongo_db"]["jobs_collection"]
-
-# Initialize MongoDB client and job store
-mongodb_client = pymongo.MongoClient(MONGODB_URL)
-mongodb_jobstore = MongoDBJobStore(
-    database=MONGODB_DATABASE,
-    collection=MONGODB_COLLECTION,
-    client=mongodb_client
-)
-
-# Initialize the scheduler with MongoDB job store
-scheduler = AsyncIOScheduler(
-    jobstores={'default': mongodb_jobstore},
-    executors={'default': AsyncIOExecutor()},
-    job_defaults={'coalesce': False, 'max_instances': 1}
-)
-
-def get_scheduler() -> AsyncIOScheduler:
-    """Get the scheduler instance."""
-    return scheduler
-
-def start_scheduler():
-    """Start the scheduler."""
-    if not scheduler.running:
-        scheduler.start()
-        print("✅ Background job scheduler started")
-
-def stop_scheduler():
-    """Stop the scheduler."""
-    if scheduler.running:
-        scheduler.shutdown()
-        print("✅ Background job scheduler stopped")
