+++ byoeb-v1/byoeb/byoeb/background_jobs/did_you_know/config.py
@@ -1,5 +1,4 @@
 import json
-from dotenv import load_dotenv
 from pathlib import Path
 
 # Get the directory of the current script
@@ -8,5 +7,4 @@ current_dir = Path(__file__).resolve().parent
 bot_config_path = current_dir / "bot_config.json"
 bot_config = json.loads(bot_config_path.read_text(encoding="utf-8"))
 
-environment_path = current_dir / ".." / ".." / ".." / "keys.env"
-load_dotenv(environment_path)
+# Note: keys.env is loaded by chat_app.configuration.config, no need to load here
