+++ byoeb-v1/byoeb/byoeb/background_jobs/message_leaderboard/leaderboard.py
@@ -4,9 +4,8 @@ from zoneinfo import ZoneInfo
 from typing import Dict, Any, List, Optional
 import pandas as pd
 
-from byoeb.background_jobs.config import app_config
-from byoeb.chat_app.configuration import dependency_setup
-from byoeb.background_jobs.dependency_setup import get_leaderboard_service, get_user_service, get_message_service
+from byoeb.chat_app.configuration.config import app_config
+from byoeb.chat_app.configuration.dependency_setup import get_leaderboard_service, user_db_service, message_db_service
 from byoeb.services.leaderboard.time_window_strategies import TimeWindowFactory
 
 IST = ZoneInfo("Asia/Kolkata")
@@ -18,8 +17,7 @@ async def fetch_phone_numbers_for_asha_and_test_users() -> List[str]:
     Returns:
         List[str]: Phone numbers of ASHA workers and test users
     """
-    user_service = await get_user_service()
-    return await user_service.fetch_phone_numbers_for_asha_and_test_users()
+    return await user_db_service.fetch_phone_numbers_for_asha_and_test_users()
 
 async def build_district_leaderboard_last_week_ist(message_categories: Optional[List[str]] = None, processing_batch_size: int = 1000) -> pd.DataFrame:
     """
@@ -113,21 +111,18 @@ async def main():
     print(f"Total recipients found: {len(phone_numbers)}")
     print(f"Message to send: {message_text}")
 
-    # Send messages using service layer
-    message_service = await get_message_service()
-
     # TEST MODE: Send only to your test phone number (COMMENTED OUT)
     test_phone_number = "917567071072"
     print(f"🧪 TEST MODE: Sending only to {test_phone_number}")
-    results = await message_service.send_bulk_messages([test_phone_number], message_text, debug_mode=False, test_mode=False)
+    results = await message_db_service.send_bulk_messages([test_phone_number], message_text, debug_mode=False, test_mode=False)
 
     # DEMO MODE: Print payloads to console without sending
     # print(f"🖥️ DEMO MODE: Printing payloads for {len(phone_numbers)} users (no actual sending)")
-    # results = await message_service.send_bulk_messages(phone_numbers, message_text, debug_mode=True)
+    # results = await message_db_service.send_bulk_messages(phone_numbers, message_text, debug_mode=True)
 
     # PRODUCTION MODE: Send to all users (actual sending) (COMMENTED OUT)
     # print(f"🚀 PRODUCTION MODE: Sending to {len(phone_numbers)} users")
-    # results = await message_service.send_bulk_messages(phone_numbers, message_text, debug_mode=False, test_mode=False)
+    # results = await message_db_service.send_bulk_messages(phone_numbers, message_text, debug_mode=False, test_mode=False)
 
     print(f"Processed {len(results)} messages via service layer")
