+++ byoeb-v1/byoeb/byoeb/services/leaderboard/strategy_example.py
@@ -5,7 +5,7 @@ import asyncio
 from byoeb.services.leaderboard.time_window_strategies import TimeWindowFactory
 from byoeb.services.leaderboard.leaderboard_service import LeaderboardService
 from byoeb.services.databases.mongo_db import UserMongoDBService, MessageMongoDBService
-from byoeb.background_jobs.dependency_setup import get_user_service, get_message_service
+from byoeb.chat_app.configuration.dependency_setup import user_db_service, message_db_service
 
 async def demonstrate_strategy_pattern():
     """Demonstrate different time window strategies for leaderboard generation."""
@@ -14,9 +14,7 @@ async def demonstrate_strategy_pattern():
     print("=" * 50)
 
     # Create services
-    user_service = await get_user_service()
-    message_service = await get_message_service()
-    leaderboard_service = LeaderboardService(user_service, message_service)
+    leaderboard_service = LeaderboardService(user_db_service, message_db_service)
 
     # Show available strategies
     print(f"📋 Available strategies: {leaderboard_service.get_available_strategies()}")
